+++ extension/src/wake.ts
@@ -1,1 +1,0 @@
-void chrome.runtime.sendMessage({ type: 'RESUME_COMPANION_WAKE' }).catch(() => undefined);