+++ plugins/resume-companion/browser-assets/extension/wake.js
@@ -1,1 +1,0 @@
-"use strict";(()=>{chrome.runtime.sendMessage({type:"RESUME_COMPANION_WAKE"}).catch(()=>{});})();